completed code ready for staging and debugging

+++ _strlen.c
@@ -6,10 +6,9 @@
  *
  * Return: length of string
  */
-
-size_t _strlen(const char *s)
+int _strlen(const char *s)
 {
-	size_t len;
+	int len;
 
 	len = 0;
 	while (*s != '\0')
